changing needle pressure source
TODO: Need the needle to stay maxed out when hose is attached and assembly body is primed. This should remain true regardless of whether the #1 shut off is open or not. Might look into checking if hose is attached && body is primed and then moving needle based on  detector values (probably zone detectors). Need to play with something along these lines.