Added Mouse input along side the already built mobile touch inputs
May end up adding a script to determine device type the user is using, then split the mouse input in to separate "PlayerController_Mouse" script to be use.
CheckDevice.cs-> PlayerController_Touch vs. PlayerControler_Mouse&Keyboard-> use one or the other.

Not sure if this is a better approach performance wise or not.