Disabled 'TipHandle' > Box Collider for Low and Bypass hoses
This 'TipHandle' is found under the hose armature.

Disabled for now due to buggy behavior with the gauge needle while two hoses are hooked up at once.

This project/ TestKitController in it's current state, is only suited for DoubleCheck testing procedures. This was by design.

In the future, if I pursue projects involving RPZ assemblies, these TipHandles will be enabled and the TestKitController will need some upgrades.